temp angular image uploader/manager
angular image uploader.  Backend still needs to be built.  Can accept
multiple images at once, checks for duplicates, and displays uploaded
images with option to remove.  Chrome can accept a folder of images,
while IE and Firefox can accept multiple selections of images but not an
actual folder... not sure if this is fixable.  Works on both of my phone
browsers. Haven't tested it in Safari, should be the same as Chrome
since they use the same webkit.